edit

+++ src/routes/Activity/components/Activity/index.js
@@ -12,37 +12,45 @@ export const Activity = (props) => {
     handleActivityUpdate,
     onActivityUpdate,
     dispatch,
-    categories
+    categories,
+    loading,
+    error
   } = props;
 
-  return (
-    <div className="activity">
-      <p>Name: {activity.name}</p>
-      <p>Description: {activity.about}</p>
-      <p>Location: {activity.location}</p>
-      <p>Date: {activity.date}</p>
+  if (loading && !(activity || {}).id) {
+    return (<div>loading</div>)
+  } else if (error) {
+    return (<p>Error!</p>);
+  } else {
+    return (
+      <div className="activity">
+        <p>Name: {activity.name}</p>
+        <p>Description: {activity.about}</p>
+        <p>Location: {activity.location}</p>
+        <p>Date: {activity.date}</p>
 
-      <ActivityInput
-        onSubmit={(props) => {
-          handleActivityUpdate({ ...props, id: activity.id }, onActivityUpdate, activity, dispatch)
-        }}
-        buttonText={"Edit Activity"}
-        placeholder={"Edit"}
-        categories={categories}
-        activity={activity}
-        activeCategoryId={activeCategoryId}
-      />
+        <ActivityInput
+          onSubmit={(props) => {
+            handleActivityUpdate({ ...props, id: activity.id }, onActivityUpdate, activity, dispatch)
+          }}
+          buttonText={"Edit Activity"}
+          placeholder={"Edit"}
+          categories={categories}
+          activity={activity}
+          activeCategoryId={activeCategoryId}
+        />
 
-      <DeleteItem
-        onSubmit={(props) => {
-          handleActivityDelete(props, onActivityDelete)
-        }}
-        id={activity.id}
-      />
-    </div>
-  );
+        <DeleteItem
+          onSubmit={(props) => {
+            handleActivityDelete(props, onActivityDelete)
+          }}
+          id={activity.id}
+        />
+      </div>
+    );
+  }
+
 };
-
 Activity.propTypes = {
   handleActivityUpdate: React.PropTypes.func.isRequired,
   onActivityUpdate: React.PropTypes.func.isRequired,
